removing unused variables and prepping for release
