+++ app/(dashboard)/_components/smart-notes/CommandMenu.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useRef, useState } from 'react';
-import { FileText, Hash, Star, Calendar, Image, LinkIcon, Lightbulb, MessageSquare, Type, ListOrdered, List } from 'lucide-react';
+import { Hash, Star, Calendar, Image, LinkIcon, Lightbulb, MessageSquare, Type, ListOrdered, List } from 'lucide-react';
 
 export interface Command {
   icon: any;
